+++ dantro/__init__.py
@@ -10,7 +10,7 @@ log = logging.getLogger(__name__)
 
 
 # Package version
-__version__ = '0.7.0'
+__version__ = '0.8.0rc1'
 
 
 # Make manager classes available