+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 `dantro` aims to adhere to [semantic versioning](https://semver.org/).
 
+## v0.8.0
+- #27 renames `ProxyMixin`-based classes to `ProxySupportMixin` to better communicate what they do and avoid confusion with `BaseDataProxy`-derived classes.
+- Also with #27, it is possible to load HDF5 data into [`dask.array`s](http://docs.dask.org/en/latest/array.html), which allow to perform lazy operations on the data. This makes it hugely more comfortable to work with large amounts of data in dantro.
+    - The `HDFDataProxy` can resolve HDF5 data as delayed `dask.array`s.
+    - The `Hdf5LoaderMixin` now allows to pass parameters to created proxies, thus allowing to create proxies which `resolve_as_dask`.
+    - The `dask.array` can be used as underlying data for the `XrDataContainer` while retaining the _exact_ same interface as with in-memory numpy data. This is possible due to the tight [integration of xarray with dask](http://xarray.pydata.org/en/stable/dask.html).
+
 ## v0.7.0
 - Infrastructure and Documentation
     - #43 Add a Sphinx-based documentation, currently containing only the API reference